nydus: move common code from nydusd into nydus crate
Move common code from nydusd into nydus crate, so it can be reused.
