+++ src/singleton.rs
@@ -2,38 +2,39 @@
 //
 // SPDX-License-Identifier: (Apache-2.0 AND BSD-3-Clause)
 
+//! Nydus daemon to host multiple services, including fscache and fusedev.
+
 use std::any::Any;
-#[cfg(target_os = "linux")]
-use std::path::Path;
 use std::sync::atomic::{AtomicBool, AtomicI32, Ordering};
 use std::sync::mpsc::{channel, Receiver, Sender};
 use std::sync::{Arc, Mutex};
 
-use nydus::blob_cache::BlobCacheMgr;
-use nydus::Result;
+use mio::Waker;
 use nydus_api::http::BlobCacheList;
 use nydus_app::BuildTimeInfo;
 
+use crate::blob_cache::BlobCacheMgr;
 use crate::daemon::{
     DaemonState, DaemonStateMachineContext, DaemonStateMachineInput, DaemonStateMachineSubscriber,
+    NydusDaemon,
 };
-use crate::{FsService, NydusDaemon, SubCmdArgs, DAEMON_CONTROLLER};
-#[cfg(target_os = "linux")]
-use nydus::validate_threads_configuration;
+use crate::fs_service::FsService;
+use crate::{Error, Result, SubCmdArgs};
 
-pub struct ServiceController {
+struct ServiceController {
     bti: BuildTimeInfo,
     id: Option<String>,
     request_sender: Arc<Mutex<Sender<DaemonStateMachineInput>>>,
     result_receiver: Mutex<Receiver<Result<()>>>,
     state: AtomicI32,
     supervisor: Option<String>,
+    waker: Arc<Waker>,
 
     blob_cache_mgr: Arc<BlobCacheMgr>,
 
     fscache_enabled: AtomicBool,
     #[cfg(target_os = "linux")]
-    fscache: Mutex<Option<Arc<nydus::fs_cache::FsCacheHandler>>>,
+    fscache: Mutex<Option<Arc<crate::fs_cache::FsCacheHandler>>>,
 }
 
 impl ServiceController {
@@ -46,13 +47,14 @@ impl ServiceController {
             if let Some(fscache) = self.fscache.lock().unwrap().clone() {
                 for _ in 0..fscache.working_threads() {
                     let fscache2 = fscache.clone();
+                    let waker = self.waker.clone();
                     std::thread::spawn(move || {
                         if let Err(e) = fscache2.run_loop() {
                             error!("Failed to run fscache service loop, {}", e);
                         }
                         // Notify the global service controller that one working thread is exiting.
-                        if let Err(e) = crate::DAEMON_CONTROLLER.waker.wake() {
-                            error!("Failed to notify the global service controller, {}", e);
+                        if let Err(err) = waker.wake() {
+                            error!("fail to exit daemon, error: {:?}", err);
                         }
                     });
                 }
@@ -75,8 +77,6 @@ impl ServiceController {
     }
 
     fn initialize_blob_cache(&self, config: &Option<serde_json::Value>) -> std::io::Result<()> {
-        DAEMON_CONTROLLER.set_blob_cache_mgr(self.blob_cache_mgr.clone());
-
         // Create blob cache objects configured by the configuration file.
         if let Some(config) = config {
             if let Some(config1) = config.as_object() {
@@ -99,7 +99,7 @@ impl ServiceController {
 impl ServiceController {
     fn initialize_fscache_service(&self, subargs: &SubCmdArgs, path: &str) -> std::io::Result<()> {
         // Validate --fscache option value is an existing directory.
-        let p = match Path::new(&path).canonicalize() {
+        let p = match std::path::Path::new(&path).canonicalize() {
             Err(e) => {
                 error!("--fscache option needs a directory to cache files");
                 return Err(e);
@@ -122,7 +122,7 @@ impl ServiceController {
         let tag = subargs.value_of("fscache-tag").map(|s| s.as_str());
 
         let threads = if let Some(threads_value) = subargs.value_of("fscache-threads") {
-            validate_threads_configuration(threads_value).map_err(|err| einval!(err))?
+            crate::validate_threads_configuration(threads_value).map_err(|err| einval!(err))?
         } else {
             1usize
         };
@@ -133,7 +133,7 @@ impl ServiceController {
             tag.unwrap_or("<none>"),
             threads
         );
-        let fscache = nydus::fs_cache::FsCacheHandler::new(
+        let fscache = crate::fs_cache::FsCacheHandler::new(
             "/dev/cachefiles",
             p,
             tag,
@@ -156,6 +156,10 @@ impl NydusDaemon for ServiceController {
         self.id.clone()
     }
 
+    fn version(&self) -> BuildTimeInfo {
+        self.bti.clone()
+    }
+
     fn get_state(&self) -> DaemonState {
         self.state.load(Ordering::Relaxed).into()
     }
@@ -164,16 +168,12 @@ impl NydusDaemon for ServiceController {
         self.state.store(state as i32, Ordering::Relaxed);
     }
 
-    fn version(&self) -> BuildTimeInfo {
-        self.bti.clone()
-    }
-
     fn start(&self) -> Result<()> {
         self.start_services()
-            .map_err(|e| nydus::Error::StartService(format!("{}", e)))
+            .map_err(|e| Error::StartService(format!("{}", e)))
     }
 
-    fn disconnect(&self) -> Result<()> {
+    fn umount(&self) -> Result<()> {
         self.stop_services();
         Ok(())
     }
@@ -187,15 +187,19 @@ impl NydusDaemon for ServiceController {
     }
 
     fn save(&self) -> Result<()> {
-        Err(nydus::Error::Unsupported)
+        Err(Error::Unsupported)
     }
 
     fn restore(&self) -> Result<()> {
-        Err(nydus::Error::Unsupported)
+        Err(Error::Unsupported)
     }
 
     fn get_default_fs_service(&self) -> Option<Arc<dyn FsService>> {
         None
+    }
+
+    fn get_blob_cache_mgr(&self) -> Option<Arc<BlobCacheMgr>> {
+        Some(self.blob_cache_mgr.clone())
     }
 
     fn delete_blob(&self, _blob_id: String) -> Result<()> {
@@ -204,10 +208,10 @@ impl NydusDaemon for ServiceController {
             if let Some(fscache) = self.fscache.lock().unwrap().clone() {
                 return fscache
                     .cull_cache(_blob_id)
-                    .map_err(|e| nydus::Error::StartService(format!("{}", e)));
+                    .map_err(|e| Error::StartService(format!("{}", e)));
             }
         }
-        Err(nydus::Error::Unsupported)
+        Err(Error::Unsupported)
     }
 }
 
@@ -217,18 +221,24 @@ impl DaemonStateMachineSubscriber for ServiceController {
             .lock()
             .unwrap()
             .send(event)
-            .map_err(|e| nydus::Error::Channel(format!("send {:?}", e)))?;
+            .map_err(|e| Error::Channel(format!("failed to send request, {}", e)))?;
+
         self.result_receiver
             .lock()
             .expect("Not expect poisoned lock!")
             .recv()
-            .map_err(|e| nydus::Error::Channel(format!("recv {:?}", e)))?
+            .map_err(|e| Error::Channel(format!("failed to receive reply, {}", e)))?
     }
 }
 
+/// Create and start a Nydus daemon to host fscache and fusedev services.
+///
+/// The `subargs` argument is derived from commandline options and controls services enabled.
+/// The fscache service will be enabled if `subargs` contains the `--fscache` argument.
 pub fn create_daemon(
     subargs: &SubCmdArgs,
     bti: BuildTimeInfo,
+    waker: Arc<Waker>,
 ) -> std::io::Result<Arc<dyn NydusDaemon>> {
     let id = subargs.value_of("id").map(|id| id.to_string());
     let supervisor = subargs.value_of("supervisor").map(|s| s.to_string());
@@ -251,6 +261,7 @@ pub fn create_daemon(
         result_receiver: Mutex::new(from_sm),
         state: Default::default(),
         supervisor,
+        waker,
 
         blob_cache_mgr: Arc::new(BlobCacheMgr::new()),
